Move BoondollarPriceRecipe and BoondollarPrices to the same package as ConsortRewardHandler
