+++ src/main/java/com/mraof/minestuck/entity/consort/BoondollarPriceRecipe.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.util;
+package com.mraof.minestuck.entity.consort;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
